+++ cartography/intel/aws/ec2/__init__.py
@@ -1,0 +1,44 @@
+import logging
+
+import botocore.config
+
+from cartography.intel.aws.ec2.auto_scaling_groups import sync_ec2_auto_scaling_groups
+from cartography.intel.aws.ec2.instances import sync_ec2_instances
+from cartography.intel.aws.ec2.key_pairs import sync_ec2_key_pairs
+from cartography.intel.aws.ec2.load_balancer_v2s import sync_load_balancer_v2s
+from cartography.intel.aws.ec2.load_balancers import sync_load_balancers
+from cartography.intel.aws.ec2.security_groups import sync_ec2_security_groupinfo
+from cartography.intel.aws.ec2.vpc import sync_vpc
+from cartography.intel.aws.ec2.vpc_peering import sync_vpc_peering
+from cartography.util import timeit
+
+logger = logging.getLogger(__name__)
+
+
+@timeit
+def get_ec2_regions(boto3_session):
+    client = boto3_session.client('ec2')
+    result = client.describe_regions()
+    return [r['RegionName'] for r in result['Regions']]
+
+
+# TODO memoize this
+def get_botocore_config():
+    return botocore.config.Config(
+        read_timeout=360,
+        retries={
+            'max_attempts': 10,
+        },
+    )
+
+
+def sync(neo4j_session, boto3_session, regions, account_id, sync_tag, common_job_parameters):
+    logger.info("Syncing EC2 for account '%s'.", account_id)
+    sync_vpc(neo4j_session, boto3_session, regions, account_id, sync_tag, common_job_parameters)
+    sync_ec2_security_groupinfo(neo4j_session, boto3_session, regions, account_id, sync_tag, common_job_parameters)
+    sync_ec2_key_pairs(neo4j_session, boto3_session, regions, account_id, sync_tag, common_job_parameters)
+    sync_ec2_instances(neo4j_session, boto3_session, regions, account_id, sync_tag, common_job_parameters)
+    sync_ec2_auto_scaling_groups(neo4j_session, boto3_session, regions, account_id, sync_tag, common_job_parameters)
+    sync_load_balancers(neo4j_session, boto3_session, regions, account_id, sync_tag, common_job_parameters)
+    sync_load_balancer_v2s(neo4j_session, boto3_session, regions, account_id, sync_tag, common_job_parameters)
+    sync_vpc_peering(neo4j_session, boto3_session, regions, account_id, sync_tag, common_job_parameters)
