+++ tests/integration/cartography/intel/aws/ec2/test_ec2_instances.py
@@ -11,7 +11,9 @@ def test_load_ec2_instances(neo4j_session, *args):
     Ensure that instances actually get loaded and have their key fields
     """
     data = tests.data.aws.ec2.instances.DESCRIBE_INSTANCES
-    cartography.intel.aws.ec2.load_ec2_instances(neo4j_session, data, TEST_REGION, TEST_ACCOUNT_ID, TEST_UPDATE_TAG)
+    cartography.intel.aws.ec2.instances.load_ec2_instances(
+        neo4j_session, data, TEST_REGION, TEST_ACCOUNT_ID, TEST_UPDATE_TAG,
+    )
 
     expected_nodes = {
         (
@@ -52,7 +54,9 @@ def test_ec2_reservations_to_instances(neo4j_session, *args):
     Ensure that instances are connected to their expected reservations
     """
     data = tests.data.aws.ec2.instances.DESCRIBE_INSTANCES
-    cartography.intel.aws.ec2.load_ec2_instances(neo4j_session, data, TEST_REGION, TEST_ACCOUNT_ID, TEST_UPDATE_TAG)
+    cartography.intel.aws.ec2.instances.load_ec2_instances(
+        neo4j_session, data, TEST_REGION, TEST_ACCOUNT_ID, TEST_UPDATE_TAG,
+    )
 
     expected_nodes = {
         (
